refactor: consolidate MutableBoardCoord type and update imports
- Moved the `MutableBoardCoord` type definition from `mutable-board-coord.ts` to `board-coords.ts` for better organization and clarity.
- Updated import statements across multiple files to reference the new location of `MutableBoardCoord`, ensuring consistency throughout the codebase.
- Removed the now redundant `mutable-board-coord.ts` file, simplifying the project structure.
