+++ src/board/chess-board.ts
@@ -1,5 +1,5 @@
 import type { BoardCoord } from '#board/board-coords';
-import type { MutableBoardCoord } from '#board/mutable-board-coord';
+import type { MutableBoardCoord } from '#board/board-coords';
 import PiecePositions from '#board/piece-positions';
 import type { Action, MoveAction, CaptureAction, PromoteAction } from '#types/actions';
 import type { ChessPiece } from '#types/game';
